UseHaskellAPIClient

Documentation

restAPI :: Proxy API

loadEnvVars :: Maybe String -> ClientM ResponseData

The following provides the implementations of these types Note that the order of the functions must match the endpoints in the type API from UseHaskell.hs

The function type of the interface here. Each function matches one of the endpoints in type API from UseHaskellAPI.hs

getREADME :: ClientM ResponseData

storeMessage :: Message -> ClientM Bool

searchMessage :: Maybe String -> ClientM [Message]

performRestCall :: Maybe String -> ClientM ResponseData